+++ admin/includes/class-post.php
@@ -123,7 +123,7 @@ class Post {
 	 * @return null
 	 */
 	public function createPost() {
-		// Fetch the post type
+		// Fetch the post's type
 		$type = $_GET['type'] ?? 'post';
 		
 		// Validate the form data and return any messages
@@ -216,10 +216,10 @@ class Post {
 								<?php
 								// Construct an image tag to display the featured image thumbnail
 								echo formTag('img', array('id'=>'media-thumb', 'src'=>'//:0', 'width'=>'100%'));
+								
+								// Construct a span tag to display the 'remove image' button
+								echo formTag('span', array('id'=>'image-remove', 'title'=>'Remove', 'content'=>formTag('i', array('class'=>'fas fa-times'))));
 								?>
-								<span id="image-remove" title="Remove">
-									<i class="fas fa-times"></i>
-								</span>
 							</div>
 							<?php
 							// Construct a hidden 'media type' form tag
@@ -392,14 +392,23 @@ class Post {
 								<div class="block">
 									<h2>Featured Image</h2>
 									<div class="row">
+										<div class="feat-image-wrap<?php echo !empty($meta['feat_image']) ? ' visible' : ''; ?>">
+											<?php
+											// Construct an image tag to display the featured image thumbnail
+											echo formTag('img', array('id'=>'media-thumb', 'src'=>getMedia($meta['feat_image']), 'width'=>'100%'));
+											
+											// Construct a span tag to display the 'remove image' button
+											echo formTag('span', array('id'=>'image-remove', 'title'=>'Remove', 'content'=>formTag('i', array('class'=>'fas fa-times'))));
+											?>
+										</div>
 										<?php
-										// Display the featured image if it's been selected
-										isset($_POST['feat_image']) && strlen($_POST['feat_image']) > 0 ? '<img src=""><span></span>' : '';
+										// Construct a hidden 'media type' form tag
+										echo formTag('input', array('type'=>'hidden', 'id'=>'media-type', 'value'=>'image'));
 										
-										// Construct a hidden 'featured' form tag
-										echo formTag('input', array('type'=>'hidden', 'name'=>'feat_image'));
+										// Construct a hidden 'featured image' form tag
+										echo formTag('input', array('type'=>'hidden', 'id'=>'media-id', 'name'=>'feat_image', 'value'=>$meta['feat_image']));
 										?>
-										<a href="#">Choose Image</a>
+										<a id="modal-launch" href="javascript:void(0)">Choose Image</a>
 									</div>
 								</div>
 							</div>
@@ -427,6 +436,8 @@ class Post {
 						</form>
 					</div>
 					<?php
+					// Include the upload modal
+					include_once PATH.ADMIN.INC.'/modal-upload.php';
 				}
 			}
 		}
@@ -560,7 +571,7 @@ class Post {
 			$data['status'] = 'draft';
 		
 		// Create an array to hold the post's metadata
-		$postmeta = array('feat_image'=>$data['feat_image'], 'title'=>$data['meta_title'], 'description'=>$data['meta_description']);
+		$postmeta = array('title'=>$data['meta_title'], 'description'=>$data['meta_description'], 'feat_image'=>$data['feat_image']);
 		
 		if($id === 0) {
 			// Set the parent to zero if the post's type is 'post' (non-hierarchical)